refactor(client): reorder methods alphabetically, move to partial files, fix interface placement
- Reorder public methods alphabetically in all client and interface files
- Move server management methods from GlideClient.cs and GlideClusterClient.cs
  into new *.ServerManagementCommands.cs partial files
- Move generic commands (Migrate, Scan, Watch, Unwatch) into new
  *.GenericCommands.cs partial files
- Move CopyAsync(destinationDatabase) and MoveAsync from IGlideClient to
  IBaseClient (both clients need them)
- Fix IGlideClusterClient.PubSubCommands.cs namespace from Valkey.Glide.Commands
  to Valkey.Glide (was incorrect partial interface split)
- Update cref references in IBatchStandalone.cs and IDatabaseAsync.GenericCommands.cs
- Add TODO #462 comments for future consolidation of redundant overrides
