+++ sources/Valkey.Glide/Abstract/IDatabaseAsync.GenericCommands.cs
@@ -321,7 +321,7 @@ public partial interface IDatabaseAsync
     /// </remarks>
     Task<bool> KeyCopyAsync(ValkeyKey sourceKey, ValkeyKey destinationKey, int destinationDatabase = -1, bool replace = false, CommandFlags flags = CommandFlags.None);
 
-    /// <inheritdoc cref="IGlideClient.MoveAsync(ValkeyKey, int)"/>
+    /// <inheritdoc cref="IBaseClient.MoveAsync(ValkeyKey, int)"/>
     /// <param name="flags">Command flags (currently not supported by GLIDE).</param>
     /// <exception cref="NotImplementedException">Thrown if <paramref name="flags"/> is not <see cref="CommandFlags.None"/>.</exception>
     Task<bool> KeyMoveAsync(ValkeyKey key, int database, CommandFlags flags = CommandFlags.None);
