+++ sources/Valkey.Glide/Client/BaseClient.ConnectionManagementCommands.cs
@@ -5,6 +5,7 @@ using Valkey.Glide.Internals;
 
 namespace Valkey.Glide;
 
+// TODO #462: Consolidate no-route overloads into BaseClient (glide-core default routing matches).
 public abstract partial class BaseClient
 {
     /// <inheritdoc cref="IBaseClient.ClientGetNameAsync()"/>
@@ -12,6 +13,18 @@ public abstract partial class BaseClient
 
     /// <inheritdoc cref="IBaseClient.ClientIdAsync()"/>
     public abstract Task<long> ClientIdAsync();
+
+    /// <inheritdoc cref="IBaseClient.ClientPauseAsync(TimeSpan)"/>
+    public abstract Task ClientPauseAsync(TimeSpan timeout);
+
+    /// <inheritdoc cref="IBaseClient.ClientPauseWriteAsync(TimeSpan)"/>
+    public abstract Task ClientPauseWriteAsync(TimeSpan timeout);
+
+    /// <inheritdoc cref="IBaseClient.ClientTrackingInfoAsync()"/>
+    public abstract Task<ClientTrackingInfo> ClientTrackingInfoAsync();
+
+    /// <inheritdoc cref="IBaseClient.ClientUnpauseAsync()"/>
+    public abstract Task ClientUnpauseAsync();
 
     /// <inheritdoc cref="IBaseClient.EchoAsync(ValkeyValue)"/>
     public abstract Task<ValkeyValue> EchoAsync(ValkeyValue message);
@@ -22,22 +35,10 @@ public abstract partial class BaseClient
     /// <inheritdoc cref="IBaseClient.PingAsync(ValkeyValue)"/>
     public abstract Task<ValkeyValue> PingAsync(ValkeyValue message);
 
-    /// <inheritdoc cref="IConnectionManagementBaseCommands.SelectAsync(long)"/>
-    public abstract Task SelectAsync(long index);
-
-    /// <inheritdoc cref="IBaseClient.ClientPauseAsync(TimeSpan)"/>
-    public abstract Task ClientPauseAsync(TimeSpan timeout);
-
-    /// <inheritdoc cref="IBaseClient.ClientPauseWriteAsync(TimeSpan)"/>
-    public abstract Task ClientPauseWriteAsync(TimeSpan timeout);
-
-    /// <inheritdoc cref="IBaseClient.ClientUnpauseAsync()"/>
-    public abstract Task ClientUnpauseAsync();
-
-    /// <inheritdoc cref="IBaseClient.ClientTrackingInfoAsync()"/>
-    public abstract Task<ClientTrackingInfo> ClientTrackingInfoAsync();
-
     /// <inheritdoc cref="IBaseClient.ResetAsync()"/>
     public async Task ResetAsync()
         => _ = await Command(Request.Reset());
+
+    /// <inheritdoc cref="IConnectionManagementBaseCommands.SelectAsync(long)"/>
+    public abstract Task SelectAsync(long index);
 }
