+++ sources/Valkey.Glide/Client/GlideClient.cs
@@ -58,6 +58,10 @@ public partial class GlideClient :
     public static async Task<GlideClient> CreateClient(StandaloneClientConfiguration config)
         => await CreateClient(config, () => new GlideClient());
 
+    /// <inheritdoc cref="IGenericCommands.CustomCommand(IEnumerable{GlideString})"/>
+    public async Task<object?> CustomCommand(IEnumerable<GlideString> args)
+        => await Command(Request.CustomCommand([.. args]));
+
     /// <inheritdoc cref="IGenericCommands.Exec(Batch, bool)"/>
     public async Task<object?[]?> Exec(Batch batch, bool raiseOnError)
         => await Batch(batch, raiseOnError);
@@ -65,183 +69,6 @@ public partial class GlideClient :
     /// <inheritdoc cref="IGenericCommands.Exec(Batch, bool, BatchOptions)"/>
     public async Task<object?[]?> Exec(Batch batch, bool raiseOnError, BatchOptions options)
         => await Batch(batch, raiseOnError, options);
-
-    /// <inheritdoc cref="IGenericCommands.CustomCommand(IEnumerable{GlideString})"/>
-    public async Task<object?> CustomCommand(IEnumerable<GlideString> args)
-        => await Command(Request.CustomCommand([.. args]));
-
-    /// <inheritdoc cref="IGlideClient.MigrateAsync(IEnumerable{ValkeyKey}, MigrateOptions)"/>
-    public async Task<bool> MigrateAsync(IEnumerable<ValkeyKey> keys, MigrateOptions options)
-        => await Command(Request.MigrateAsync(keys, options));
-
-    /// <inheritdoc cref="IGlideClient.InfoAsync()"/>
-    public async Task<string> InfoAsync() => await InfoAsync([]);
-
-    /// <inheritdoc cref="IGlideClient.InfoAsync(IEnumerable{InfoOptions.Section})"/>
-    public async Task<string> InfoAsync(IEnumerable<InfoOptions.Section> sections)
-        => await Command(Request.Info([.. sections]));
-
-    /// <inheritdoc cref="IGlideClient.ConfigGetAsync(ValkeyValue)"/>
-    public async Task<KeyValuePair<string, string>[]> ConfigGetAsync(ValkeyValue pattern = default)
-        => await Command(Request.ConfigGetAsync(pattern));
-
-    /// <inheritdoc cref="IGlideClient.ConfigResetStatisticsAsync()"/>
-    public async Task ConfigResetStatisticsAsync()
-        => _ = await Command(Request.ConfigResetStatisticsAsync());
-
-    /// <inheritdoc cref="IGlideClient.ConfigRewriteAsync()"/>
-    public async Task ConfigRewriteAsync()
-        => _ = await Command(Request.ConfigRewriteAsync());
-
-    /// <inheritdoc cref="IGlideClient.ConfigSetAsync(ValkeyValue, ValkeyValue)"/>
-    public async Task ConfigSetAsync(ValkeyValue setting, ValkeyValue value)
-        => _ = await Command(Request.ConfigSetAsync(setting, value));
-
-    /// <inheritdoc cref="IBaseClient.ConfigSetAsync(IDictionary{ValkeyValue, ValkeyValue})"/>
-    public override async Task ConfigSetAsync(IDictionary<ValkeyValue, ValkeyValue> parameters)
-        => _ = await Command(Request.ConfigSetAsync(parameters));
-
-    /// <inheritdoc cref="IGlideClient.DatabaseSizeAsync()"/>
-    public async Task<long> DatabaseSizeAsync()
-        => await Command(Request.DatabaseSizeAsync());
-
-    /// <inheritdoc cref="IGlideClient.FlushAllDatabasesAsync()"/>
-    public async Task FlushAllDatabasesAsync()
-        => _ = await Command(Request.FlushAllDatabasesAsync());
-
-    /// <inheritdoc cref="IBaseClient.FlushAllDatabasesAsync(FlushMode)"/>
-    public override async Task FlushAllDatabasesAsync(FlushMode mode)
-        => _ = await Command(Request.FlushAllDatabasesAsync(mode));
-
-    /// <inheritdoc cref="IGlideClient.FlushDatabaseAsync()"/>
-    public async Task FlushDatabaseAsync()
-        => _ = await Command(Request.FlushDatabaseAsync());
-
-    /// <inheritdoc cref="IBaseClient.FlushDatabaseAsync(FlushMode)"/>
-    public override async Task FlushDatabaseAsync(FlushMode mode)
-        => _ = await Command(Request.FlushDatabaseAsync(mode));
-
-    /// <inheritdoc cref="IBaseClient.SaveAsync()"/>
-    public override async Task SaveAsync()
-        => _ = await Command(Request.SaveAsync());
-
-    /// <inheritdoc cref="IGlideClient.BgRewriteAofAsync()"/>
-    public async Task<string> BgRewriteAofAsync()
-        => await Command(Request.BgRewriteAofAsync());
-
-    /// <inheritdoc cref="IGlideClient.LastSaveAsync()"/>
-    public Task<DateTimeOffset> LastSaveAsync()
-        => Command(Request.LastSaveAsync());
-
-    /// <inheritdoc cref="IGlideClient.TimeAsync()"/>
-    public Task<DateTimeOffset> TimeAsync()
-        => Command(Request.TimeAsync());
-
-    /// <inheritdoc cref="IGlideClient.LolwutAsync()"/>
-    public async Task<string> LolwutAsync()
-        => await Command(Request.LolwutAsync());
-
-    /// <inheritdoc cref="IBaseClient.LolwutAsync(LolwutOptions)"/>
-    public override async Task<string> LolwutAsync(LolwutOptions options)
-        => await Command(Request.LolwutAsync(options));
-
-    /// <inheritdoc cref="IGlideClient.MemoryDoctorAsync()"/>
-    public async Task<string> MemoryDoctorAsync()
-        => await Command(Request.MemoryDoctorAsync());
-
-    /// <inheritdoc cref="IGlideClient.MemoryMallocStatsAsync()"/>
-    public async Task<string> MemoryMallocStatsAsync()
-        => await Command(Request.MemoryMallocStatsAsync());
-
-    /// <inheritdoc cref="IGlideClient.MemoryPurgeAsync()"/>
-    public async Task MemoryPurgeAsync()
-        => _ = await Command(Request.MemoryPurgeAsync());
-
-    /// <inheritdoc cref="IGlideClient.MemoryStatsAsync()"/>
-    public async Task<MemoryStats> MemoryStatsAsync()
-        => await Command(Request.MemoryStatsAsync());
-
-    /// <inheritdoc cref="IBaseClient.ConfigGetAsync(IEnumerable{ValkeyValue})"/>
-    public override async Task<KeyValuePair<string, string>[]> ConfigGetAsync(IEnumerable<ValkeyValue> patterns)
-        => await Command(Request.ConfigGetAsync(patterns));
-
-    /// <inheritdoc cref="IGlideClient.BackgroundSaveAsync()"/>
-    public async Task<string> BackgroundSaveAsync()
-        => await Command(Request.BackgroundSaveAsync());
-
-    /// <inheritdoc cref="IGlideClient.BackgroundSaveScheduleAsync()"/>
-    public async Task<string> BackgroundSaveScheduleAsync()
-        => await Command(Request.BackgroundSaveScheduleAsync());
-
-    /// <inheritdoc cref="IGlideClient.BackgroundSaveCancelAsync()"/>
-    public async Task<string> BackgroundSaveCancelAsync()
-        => await Command(Request.BackgroundSaveCancelAsync());
-
-    /// <inheritdoc cref="IGlideClient.FailoverAsync()"/>
-    public async Task FailoverAsync()
-        => _ = await Command(Request.FailoverAsync());
-
-    /// <inheritdoc cref="IGlideClient.FailoverAsync(FailoverOptions)"/>
-    public async Task FailoverAsync(FailoverOptions options)
-        => _ = await Command(Request.FailoverAsync(options));
-
-    /// <inheritdoc cref="IGlideClient.ReplicaOfAsync(string, int)"/>
-    public async Task ReplicaOfAsync(string host, int port)
-        => _ = await Command(Request.ReplicaOfAsync(host, port));
-
-    /// <inheritdoc cref="IGlideClient.ReplicaOfNoOneAsync()"/>
-    public async Task ReplicaOfNoOneAsync()
-        => _ = await Command(Request.ReplicaOfNoOneAsync());
-
-    /// <inheritdoc/>
-    public async Task<(string cursor, ValkeyKey[] keys)> ScanAsync(string cursor, ScanOptions? options = null)
-        => await Command(Request.ScanAsync(cursor, options));
-
-    /// <inheritdoc cref="IGlideClient.ScanAsync(ScanOptions?)"/>
-    public async IAsyncEnumerable<ValkeyKey> ScanAsync(ScanOptions? options = null)
-    {
-        string currentCursor = "0";
-
-        do
-        {
-            (string nextCursor, ValkeyKey[] keys) = await Command(Request.ScanAsync(currentCursor, options));
-
-            foreach (ValkeyKey key in keys)
-            {
-                yield return key;
-            }
-
-            currentCursor = nextCursor;
-        } while (currentCursor != "0");
-    }
-
-    /// <inheritdoc cref="ITransactionBaseCommands.WatchAsync(IEnumerable{ValkeyKey})"/>
-    public async Task WatchAsync(IEnumerable<ValkeyKey> keys)
-        => _ = await Command(Request.Watch(keys));
-
-    /// <inheritdoc cref="ITransactionCommands.UnwatchAsync()"/>
-    public async Task UnwatchAsync()
-        => _ = await Command(Request.Unwatch());
-
-    /// <inheritdoc cref="IGlideClient.LatencyHistoryAsync(ValkeyValue)"/>
-    public async Task<LatencyEntry[]> LatencyHistoryAsync(ValkeyValue @event)
-        => await Command(Request.LatencyHistoryAsync(@event));
-
-    /// <inheritdoc cref="IGlideClient.LatencyLatestAsync()"/>
-    public async Task<LatencyEventInfo[]> LatencyLatestAsync()
-        => await Command(Request.LatencyLatestAsync());
-
-    /// <inheritdoc cref="IBaseClient.LatencyResetAsync()"/>
-    public override async Task<long> LatencyResetAsync()
-        => await Command(Request.LatencyResetAsync([]));
-
-    /// <inheritdoc cref="IBaseClient.LatencyResetAsync(ValkeyValue)"/>
-    public override async Task<long> LatencyResetAsync(ValkeyValue @event)
-        => await Command(Request.LatencyResetAsync([@event]));
-
-    /// <inheritdoc cref="IBaseClient.LatencyResetAsync(IEnumerable{ValkeyValue})"/>
-    public override async Task<long> LatencyResetAsync(IEnumerable<ValkeyValue> events)
-        => await Command(Request.LatencyResetAsync(events));
 
     /// <inheritdoc cref="BaseClient.GetServerVersionAsync()"/>
     protected override async Task<Version> GetServerVersionAsync()
