+++ sources/Valkey.Glide/Client/IGlideClient.cs
@@ -18,47 +18,6 @@ public partial interface IGlideClient :
     IServerManagementStandaloneCommands
 {
     /// <summary>
-    /// Moves key from the currently selected database to the specified destination database.
-    /// When key already exists in the destination database, or it does not exist in the source database, it does nothing.
-    /// It is possible to use MOVE as a locking primitive because of this.
-    /// </summary>
-    /// <seealso href="https://valkey.io/commands/move"/>
-    /// <param name="key">The key to move.</param>
-    /// <param name="database">The database to move the key to.</param>
-    /// <returns><see langword="true"/> if key was moved. <see langword="false"/> if key was not moved.</returns>
-    /// <remarks>
-    /// <example>
-    /// <code>
-    /// ValkeyKey key = "mykey";
-    /// bool result = await client.MoveAsync(key, 2);
-    /// </code>
-    /// </example>
-    /// </remarks>
-    Task<bool> MoveAsync(ValkeyKey key, int database);
-
-    /// <summary>
-    /// Copies the value stored at the source to the destination key in the specified database. When
-    /// replace is true, removes the destination key first if it already
-    /// exists, otherwise performs no action.
-    /// </summary>
-    /// <seealso href="https://valkey.io/commands/copy"/>
-    /// <param name="source">The key to the source value.</param>
-    /// <param name="destination">The key where the value should be copied to.</param>
-    /// <param name="destinationDatabase">The database ID to store destination in.</param>
-    /// <param name="replace">Whether to overwrite an existing values at destination.</param>
-    /// <returns><see langword="true"/> if source was copied. <see langword="false"/> if source was not copied.</returns>
-    /// <remarks>
-    /// <example>
-    /// <code>
-    /// ValkeyKey source = "source";
-    /// ValkeyKey dest = "dest";
-    /// bool result = await client.CopyAsync(source, dest, 1, replace: true);
-    /// </code>
-    /// </example>
-    /// </remarks>
-    Task<bool> CopyAsync(ValkeyKey source, ValkeyKey destination, int destinationDatabase, bool replace = false);
-
-    /// <summary>
     /// Atomically transfers keys from the source instance to the destination instance.
     /// </summary>
     /// <seealso href="https://valkey.io/commands/migrate/">Valkey commands – MIGRATE</seealso>
