+++ sources/Valkey.Glide/Client/IGlideClusterClient.PubSubCommands.cs
@@ -1,6 +1,8 @@
 // Copyright Valkey GLIDE Project Contributors - SPDX Identifier: Apache-2.0
 
-namespace Valkey.Glide.Commands;
+using Valkey.Glide.Commands;
+
+namespace Valkey.Glide;
 
 // ATTENTION: Methods should only be added to this interface if they are implemented
 // by Valkey GLIDE clients but NOT by StackExchange.Redis databases. Methods implemented
